chore: handle eruption jobs
